+++ src/plugin.hpp
@@ -16,4 +16,5 @@ extern Model *modelInfix;
 extern Model *modelStrip;
 extern Model *modelEightFace;
 extern Model *modelMidiCat;
+extern Model *modelSipo;
 extern Model *modelAudioInterface64;